% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmean_function.R
\name{tmean}
\alias{tmean}
\title{Trimmed Mean}
\usage{
tmean(x, s, l)
}
\arguments{
\item{x}{= vector
s = number of smaller values to ignore
l = number of larger values to ignore}
}
\description{
This function allows you to express your love of cats.
}
\examples{
tmean()
}
\keyword{mean,}
\keyword{trimmed}
